style: Adjusted modal container for Withdraw modal, fixed background scroll
